@@ -3,20 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@mui/material": "^5.8.2",
-    "@mui/icons-material": "^5.8.2",
-    "@emotion/styled": "^11.8.1",
-    "@emotion/react": "^11.9.0",
+    "@mui/material": "^5.14.6",
+    "@mui/icons-material": "^5.14.6",
+    "@emotion/styled": "^11.11.0",
+    "@emotion/react": "^11.11.1",
     "i": "^0.3.7",
-    "npm": "^8.10.0",
-    "react": "^18.1.0",
-    "react-dom": "^18.1.0",
-    "react-minimal-pie-chart": "^5.0.0",
-    "react-redux": "^7.1.0",
-    "react-router": "^5.0.1",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "3.1.1",
-    "redux": "^4.0.4"
+    "npm": "^9.8.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-minimal-pie-chart": "^8.4.0",
+    "react-redux": "^8.1.2",
+    "react-router": "^6.15.0",
+    "react-router-dom": "^6.15.0",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.1"
   },
   "scripts": {
     "start": "react-scripts start",
