@@ -1,5 +1,5 @@
 import React from 'react'
-import { Switch, Route } from 'react-router'
+import { Routes, Route} from 'react-router-dom'
 import Home from './containers/Home'
 import About from './components/About'
 import Car from './containers/Car'
@@ -7,12 +7,12 @@ import Dashboard from './containers/Dashboard'
 
 const Router = () => {
     return (
-        <Switch>
-            <Route exact path="/" component={Home} />
-            <Route path="/about" component={About} />
-            <Route path="/car/:id" component={Car} />
-            <Route path="/dashboard" component={Dashboard} />
-        </Switch>
+        <Routes>
+            <Route path="/" element={<Home />} />
+            <Route path="/about" element={<About/>} />
+            <Route path="/car/:id" element={<Car/>} />
+            <Route path="/dashboard" element={<Dashboard/>} />
+        </Routes>
     );
 };
 
