@@ -1,18 +1,24 @@
 import React from 'react'
-import PieChart from 'react-minimal-pie-chart';
+import { PieChart } from 'react-minimal-pie-chart';
 
 const Chart = (props) => {
-    // create variable "over" with all the cars whose horsepower is >= 200
+    if (!props.cars || !Array.isArray(props.cars)) {
+        console.log("not a car:", props.cars);
+        return <div>no data available</div>
+    }
+    // create variable "over" with all the c((ars whose horsepower is >= 200
+    const over =  props.cars.filter(car => car && car.horsepower >=200)
     // create variable "under" with all the cars whose horsepower is < 200
-
+    const under = props.cars.filter(car => car && car.horsepower < 200)
     return (
         <div>
             <PieChart style={{ width: '200px' }}
                 data={[
-                    { title: 'Over', value: 10, color: '#C13C37' },
-                    { title: 'Under', value: 15, color: '#E38627' },
+                    { title: 'Over', value: over.length, color: '#C13C37' },
+                    { title: 'Under', value: under.length, color: '#E38627' },
                 ]}
-                label
+                label={({ dataEntry }) => `${dataEntry.title} : ${dataEntry.value}`}
+                
                 labelStyle={{
                     fill: 'white',
                     fontSize: 'small'
