@@ -1,6 +1,7 @@
 import { connect } from 'react-redux'
 import Dashboard from '../components/Dashboard'
 // import { removeCar } action here
+import { removeCar } from '../redux/action'
 
 const mapStateToProps = (state) => {
     return {
@@ -10,5 +11,8 @@ const mapStateToProps = (state) => {
 }
 
 // add mapDispatchToProps function here
+const mapDispatchToProps = (dispatch) => ({
+    removeCar: (index) => dispatch(removeCar(index))
+})
 
-export default connect(mapStateToProps)(Dashboard)
+export default connect(mapStateToProps, mapDispatchToProps)(Dashboard)
